requirements.txt
streamlit==1.31.0
vapi_python
pyaudio
python-dotenv==1.0.0
